STY: Rename file cutoffFunctions to cutoff_functions
